(Binary Simulator) Clone detection avoids sqlite3 locking errors
When performing semantic analysis in parallel, the simulator obtains
write locks during certain database operations. The sqlite3x API did
not try hard enough to obtain a write lock when the database was
already locked--it just threw an exception immediately.  We now
block for up to 10 seconds when trying to obtain a lock.