(Policies) Use Git info when available
Some of the policies don't work correctly when run in a source tree
where lots of temporary developer files are also sitting.  These
temporary files are included in the list of files that are checked
by the policies.  Therefore, the FileLister has been modified to
use Git to determine whether a file is part of ROSE or a temporary
file when run inside a Git repository.