(SQLite3) Changed lock polling from 1 second to 0.1 second
When a database is locked, sqlite3x polls the lock once every
0.1 second (was 1 second) until the lock is available or until
we've waited too long.  Polling once per second resulted in
too much idle time, so we now poll 10 times per second.

Clone detection in  projects/simulator running in 10 processes
on a 12-core system was able to utilize about 50% of the system,
but now utilizes about 80%.  This is on an sqlite database that's
about 70 MB.