(Binary Analysis) Able to generate generic names for missing registers
The RegisterDictionary::lookup() method could only be used to convert
a RegisterDescriptor to a name string when a register dictionary was
available, and it bailed out with an empty string if the descriptor
did not exist in the dictionary.

A new class, RegisterNames, is able to handle cases where a register
dictionary isn't available or where the dictionary doesn't contain
an entry for the specified RegisterDescriptor. Also, this class allows
customization of how the name is generated.