(Binary Analysis) Symbolic simplification
Most of ROSE's symbolic simplification was performed in the instruction
semantics RiscOperators, but they had two problems: (1) the scope of the
simplifications was limited to the new exression node being created in
the Risc operator, and (2) the simplifications were not available when
expressions were built outside the RiscOperators.  This commit fixes both
of those problems.