(Binary Analysis) Moved byte order stuff to its own namespace
Stuff related to byte order was originally part of the SgAsmGenericFileFormat
class, which made it difficult to use in parts of ROSE that aren't
related to parsing a binary container for holding an executable, library,
object file, or coredump.

So I moved all this stuff into a ByteOrder namespace. Other than the
name space, only one name changed: the SgAsmGenericFileFormat::ByteOrder
enum is renamed to ByteOrder::Endianness.