(Binary Simulator) Clone detection outputs handled differently
The long-running part of the semantic analysis generates as much
detailed info about outputs as possible and stores it all in a
database.  A separate tool, compare_outputs, was created that
compares outputs to decide when two functions should be considered
to have produced similar outputs.  The compare_outputs is run
before the SQL that generages the list of clone pairs.