(CMAKE ROSE) more progress on CompileTests
This commit adds the the compile tests from the following
directories:
  C_tests, Cxx_tests, C_subset_of_Cxx_tests, and UPC_tests.

With a couple of exceptions, all of these tests pass.  In order to
get the UPC_tests passing, I had to modify the CMake procedure that
generates the include-staging directory.  This now copies upc.h
and other related header files into the include-staging directory,
matching the behavior of the autoconf build system.