(Support) Bug-workaround for libpqxx
The PostgreSQL library seems to have a bug when a single process opens
two connections to the same database at the same time, creates
two transactions to exist concurrently, runs a failing command
in one transaction, then deletes all objects.  libpqxx gets a
segfault in the std::map destructor for one of the connections.

This commit changes SqlDatabase in order to work around the bug at
the cost of leaking memory for the connections with the failed transaction.