(CMAKE ROSE) add Fortran compile tests
Convert the compile tests in the Fortran_tests directory to run
using CTest.

When enable-fortran option is turned on, CMake now checks whether
or not gfortran was found.  This information is used to properly
define the symbol USE_GFORTRAN_IN_ROSE.  If gfortran is found,
CMake then queries its version, as this information is needed
to know which tests are expected to pass.
There is still some work to do in the subdirectories of Fortran_tests.

Conflicts:
	CMakeLists.txt