(Clone Detection) Removed clustering tool
The 32-clusters-from-pairs tool was removed (along with its table
in the schema) because it was based on a flawed algorithm that
I found on-line.  We might replace it later with a better
implementation.

The computation for failure rates (false negatives and false positives)
was changed to avoid using clusters and instead to use the function
similarity pairs directly.  This step is now somewhat slower.

The failure rates can be calculated now based on only those pairs
of functions for which both functions pass for at least one test
in common. This makes the rates better because it avoids the
false negatives that resulted from us not being able to choose
any good inputs for the function pair.