(Clone Detection) Bug fix for 25-run-tests for multiple specimens
The code that recomputes interpretation-specific information is
triggered by a change in the address of the SgAsmInterpretation node.
However, it failed to account for the fact that if the AST is
deleted and a new AST is created, the new SgAsmInterpretation will
likely have the same address as the old node.