(Clone Detection) Changed how built-in functions are handled
Built-in functions were originally implemented within the analysis
in terms of hard-coded calls to the instruction semantics.  Although
this was fast, it was overly complex in order to handle timeouts,
and bad inputs that could cause long running loops and/or generate
large numbers of memory cells.

The new approach is to write the built-in functions in C, compile
them, and dynamically link them into the executable during the
analysis.