(EDG) Validate binary compatibility signature before download
If the computed EDG binary compatibility signature is
not valid, report an error message and fail immediately,
rather than waiting for wget to fail with a reason that
is not clear to the untrained eye.

Also, remove the lazy step of copying tarballs from within
the Lab instead of using wget. This step only adds a level
of complexity that is confusing to users and does not allow
us to properly test compiling ROSE in a way that our external
users are using.