(program transformation): Added APIs to query the status of normalization and made it more safe.
Changed extractFunctionArgumentsNormalization by adding more APIs to query the status of transformation.
IsNormalized tell if the subtree is in normalized form.
IsNormalizable tells if the given tree can be fully normalized by our pass.

Identified a safety issue in short circuit expressions. Now we don't normalize function calls that are on the RHS of any
short circuit expression.

Also, added a flag to NormalizeTree(SgNode* tree, bool doUnsafeNormalization = false). If doUnsafeNormalization is true, we lift
all function calls even if they are unsafe.