(Tests) Fix Makefile bug causing redundant make-check
As of b9fe5435a, C++ test codes include file paths in their
file names:

  ```Makefile
  ROSE__CXX_TESTS = /path/to/cxx/test.cpp
  ```

This confuses Make because the corresponding object files,
and therefore the Make targets, have paths in their names,
but the generated object files are not located at the path
specified in the name of the target:

   ```Makefile
   ROSE__CXX_OBJECTS = ${ROSE__CXX_TESTS:.cpp=.o}

   $(ROSE__CXX_OBJECTS):
       <compile>
   ```

The `<compile>` step generates the object file in the
current directory, not at "/path/to/cxx/". So Make will
always run the targets `$(ROSE__CXX_OBJECTS)` since it
thinks they are all out-of-date (non-existant).