(Tests) Remove path from $(ROSE__CXX_TESTS) as new default
Rather than include the complete file path for each C++
test code in $(ROSE__CXX_TESTS), we force users to utilize
$(ROSE__CXX_TESTS_DIR) in conjunction with the actual list
of test codes specified in $(ROSE__CXX_TESTS).

This new usage is necessary to simplify the process of using
the $(ROSE__CXX_TESTS), the global list of test codes.

-----
TIP: A section of code worth pointing out:

    ```Makefile
    $(TEST_Objects): %.o: $(ROSE__CXX_TESTS_DIR)/%.C
    ```

The `%` is called a "% Stem" and essentially acts as a wild
card match. An example will best illustrate its use:

   ```Makefile
   hello.o: %.o: /my/src/dir/path/%.C
   ```

`%` will match "hello" and therefore this line of code
would expand to:

   ```Makefile
   hello.o: %.o: /my/src/dir/path/hello.C
   ```