(Tests) Replace $(builddir) with explicit "./" path
Older versions of Autoconf do not automatically define
the Makefile variable `$(builddir)`, therefore this
variable will expand to the empty string:

    "$(builddir)/$$(basename "$@")"

Expands to:

    "/$$(basename "$@")"

Resulting in a "Permission Denied" error due to the
fact that "/" is a root system path.

The fix is to simply remove the usage of $(builddir):

    "./$$(basename "$@")"

----
Also removed unncessary copying of source code file to
build directory.