(Clone Detection) fio table tracks reasons for input consumption
The old columns tracked the number of values consumed from particular queues. The
new semantics are that those columns track the number of times an input was consumed
for a particular reason.  The two can be different when input queues are redirected.

For instance, if a function initializes five function arguments by reading those
argument memory locations, but the argument queue is redirected to the integer queue,
the old semantics were that the integer arguments_consumed column was zero and the
integers_consumed column was five.  The new semantics store it the opposite
way.