(Binary Analysis) Fixed RegisterStateGeneric::zero()
Initializes registers to zero. Old behavior was to clear the list
of register parts and set a Boolean that caused registers to be
initialized to zero when they "spring into existence" when read for
the first time.  Now we can distinguish between "register initialized
to zero" and "register not initialized".

BTW, zero() uses the set of largest non-overlapping registers, but
I don't think it matters since reading from a smaller register will
do an extract() on a concrete value, which I think most semantic
domains will constant fold without any trouble.