(Binary Analysis) Backported semantics formatting from API2 to API1
Some of the BaseSemantics::Formatter stuff has been back-ported from
the instruction semantics API2 to the old API1.  The API2 method uses
polymorphic classes rather than templates.  The goal is to be able to
write code that looks more similar between API1 and API2.

Not all properties of the API2 Formatter are applicable to API1, so
only those that are necessary were copied.