(Binary Analysis) Unparser does better job of disassembling static data
The static data disassembly was dependent on the disassembler configuration
and did things like skip over bytes that resulted in a basic block that
included an invalid instruction, disassembling some bytes twice,  etc.

The new static data disassembler just starts at the first byte
and disassembles one instruction at a time, advancing to either the
address following the instruction (on success) or the next address
(on failure).