(Binary Analysis) Disassembler tool can pre-reserve addresses
Added a --reserve switch to the "disassemble" so that regions
of the virtual address space can be reserved before the tool tries
to load the specimen and its libraries.  One use for this is to
force a dynamic object specimen to be loaded to a higher address
than the default, which is zero. E.g.,

$ ./disassemble --reserve=0,0x08000000 libfoo.so

causes the library to load at 0x08000000 because all addresses
below that have been taken.