(Clone Detection) Use BinaryLoader to relocate dynamic object specimens
This commit replaces the hack that relocates a specimen which is a
dynamic object (*.so). The old method was to change the DO's base
virtual address. The new method temporarily allocates a restricted
area in the low part of the address space, then calls the BinaryLoader
to map the DO into memory, then releases the low addresses.  This
has the benefit of also applying relocation fixups.