(KeepGoing) Remove redundant assertion that causes failure
Assertion is not needed, as the case is handled by the
proceeding conditional:

    ```C++
    if (original_file.string() != unparsed_file.string())
    ```

Moreover, this assertion will always fail in the case
where the unparsed file fails compilation, and compileOutput
is called again on the original input file. This is because,
before calling compileOutput the second time, we set the
unparsed file name to actually be the original file name:

    ```C++
    set_unparse_output_filename(outputFilename);
    ```

Therefore, the two strings will be equal.