(Binary Analysis) Fixed a bug in the new Partitioner::disassembleInterpretation()
This method should not have assumed that a memory map was available
in the interpretation. If there is no memory map, then we need to
perform the BinaryLoader::remap() step.

Also, we now run the disassembler first, then the partitioner instead
of only the partitioner driving the disassembler.  This allows the
partitioner to find functions by pattern matching instructions.