(KeepGoing) Add signal handling to CLI and Backend Compiler processing
Add two new signal handling macros:

    KEEP_GOING_CAUGHT_FRONTEND_COMMANDLINE_SIGNAL
    KEEP_GOING_CAUGHT_BACKEND_COMPILER_SIGNAL

In the case of a commandline processing failure, the only thing
that we can do is halt the program--it is unrecoverable. It is not
possible to set any SgProject attributes since the CLI processing
is currently invoked within the SgProject constructor, therefore
at the point of failure, the SgProject is not a valid object.

The backend compiler signal handling is added specifically to
address failures that are separate from the backend unparsing.
Otherwise, KeepGoing ends up in an infinite loop when there is
a failure in the backend compiler because the signals are handled
by the unparser's signal handling.