(Binary Analysis) Bug fix for thunk recognition
When analyzing a function to find a thunk, it is sometimes
possible for the initial JMP to have a target that is outside the
function. This can happen when the function contains other blocks
that are not reachable, but were added to the functionanyway
because they look statistically like code.

Conflicts:
	src/frontend/Disassemblers/Partitioner.C
		Independently fixed on another branch, and
		using that version.  The only thing we're
		keeping on this branch is an assertion.