(Clone Detection) Tested function should not be considered a called function
When testing function F, the top-level invocation of F by 25-run-tests
should not be considered a function call since there isn't actually
a call.  This change improves the accuracy of the function-returns-a-value
analysis and has a large beneficial effect on false positives without
much of an effect on false negatives.