(Snippets) Fixed bug related to include directives and global variables
Since C preprocessor directives are attached to other AST nodes, whenever
we copied something from the snippet AST into the application AST, we
also copied the directives along with it.  That meant sometimes copying
an #include directive along with the global variable that followed it even
though the #include wasn't actually needed.

This commit changes the behavior so that #include directives are always
removed from the root node of any AST copied from the snippet file.