Added Snippet::getEnclosingFileNode
The SageInterface::getEnclosingFileNode doesn't work for Java because
SgFile is not a parent of most Java nodes like it is for other
languages.  Therefore, we need to do a full traversal to find the file.

I'm not adding this to SageInterface yet because it doesn't seem like
the best algorithm.