(Build) Use absolute paths for $(srcdir), $(builddir), etc.
Normally when the user invokes "configure" with a relative path the
$(srcdir), $(builddir), $(top_srcdir), and $(top_builddir) makefile
variables are relative also. This causes problems when command-lines
are built in the makefile and passed to some other program that runs
them in a different directory (which we do for parallelism).  The
solution is to cause config.status to filter all the Makefile so that
those four variables are defined in terms of their "abs_" counterparts,
like:
  srcdir = $(abs_srcdir)
etc.