(Snippets) Java tests don't move local var declarations
Since Java allows declarations and non-declarations to be mixed
in any order within a scope, it is pointless to move declarations
to the beginning of the scope where the snippet is inserted. The
default is to leave Java declarations in their relative order
with respect to non-declarations, although moving them is still
supported if desired.