(TALCDataLayout) Automake *_SOURCE variables should be relative names
Some of the makefiles for this project had rules that had inter-directory
dependencies, but the directories were specified in such a ways as to
hard-code the project to a specific directory (projects/TALCDataLayout)
within the ROSE source tree.  A project's makefiles should be written
in such a way that we can easily move the project to a different directory
and it sill just works.

Incidentally, this also works around a problem with GNU automake.