(C Tests) Including <linux/types.h> is not portable
Applications should never include anything from /usr/include/linux if
they expect to be portable across multiple versions of Linux.

The C99 fixed-length integer types are defined in <inttypes.h> and
<stdint.h>, so one of them should be included.

The "u_char" type is not a standard type and using these ugly types
is not portable because there's no way to tell whether we need to
define the type ourselves, and defining it ourself when it is already
defined is an error.  I've replaced "u_char" with "unsigned char".