(C++ Tests) Any test specimen that includes boost headers must have boost include directories
The extractFunctionArgumentsTests assumed that boost header files are located in a well
known directory, but this isn't always the case.  I've changed the test to use the same
boost header directory as used by librose itself (although no attempt to add it earliest
in the search path, so the headers in /usr/include might still get used instead of the
boost headers used by librose).