(Snippets) Avoid calling SageInterface::findOrInsertJavaPackage for "" package.
findOrInsertJavaPackage fails when the package name is "". The failure
is a syntax error before the semicolon in the generated code "package ;"

Therefore, when the package name is empty (i.e., no "package" statement
in the snippet) we avoid calling findOrInsertJavaPackage and just hope
that the "" package was already defined.

This bug has been reported to Philippe.