(Testing) Rewrote a very messy makefile while trying to debug what passes
The various lists of test specimens were confusing because it wasn't clear
which lists were actually used and whether commenting out a list meant that
the test for those specimens fail, pass, or are not run.

The new makefile runs the exact same tests for "make check" as the old
makefile, with the exception that test2013_242.C is now marked as a failing
test and not run.  This specimen fails with an error "(problem with Boost
code in ROSE compiling ROSE)" and the assertion "current_scope->lookup_symbol(
aliasedName,__null,__null) != __null" failed. This is on Debian Squeeze
using x86_64-linux-gnu g++ 4.4.5-8 and boost 1.47.