(Snippets) Snippet tests check that injection actually occurred
Previously these tests only checked whether the output could be
compiled.  Now they also look for a particular string in the
output that could only have come from an injection.

Also removed the >132 column lines from the makefile.  These
were just comments that showed an old command-line that might
not even be valid.  Look in the *.passed/*.failed files for
the actual command-lines.