(Binary Analysis) Analysis to find post-padding function entry points
This attempts to improve upon the analysis that finds functions between
padding.  The new analysis iteratively looks for padding and discovers a
function rather than first looking for all padding and then discovering
all functions between padding.  The goal is to avoid splitting one
function that happens to contain something that looks like padding but
isn't into two or more functions.

This analysis is currently commented out near line 3525 of Partitioner.C
because it's too slow. Fixing the slowness issue will require changing
the way we keep track of used/unused bytes in the address space.