(Binary Analysis) Removed a SymbolicSemantics::SValue constructor
I removed the SymbolicSemantics::SValue constructor that takes an InsnSemanticsExpr
as an argument.  Instead, use one of the other constructors and then call set_expression.

This is removed because when an SValue class defines new constructors it places more
responsibilities on all the subclasses.  Each subclass must define the same static
allocating constructors or else the wrong type gets allocated.