(Support) Added includes for Sawyer files
These headers aren't actually used in this branch but are used in other branches
that have not been merged.  The purpose of this branch is only to introduce the
Sawyer library so that other branches will be able to use it.  Therefore, I've
added a couple extraneous #include for Sawyer headers in the places they're used
in other branches in order to bypass the policy that all headers must be used.