(Edit Distance) Added demo for binary edit distance
A separate program is needed because the purpose of frontend(), at
least for binaries, is to parse a coherent collection of specimens
whose parts create a whole process. Therefore we need to call
the equivalent of frontend() twice in order to parse two unrelated
specimens and create two processes that can be compared.