Restructured the unit testing integration.
Created the new unitTesting directory in roseTests.
Moved the gtest sources to a subdirectory of that directory.
Moved the existing unit tests as subdirectories of that directory.
Adapted the config/Makefiles.
Added the testSupport header file, where convenience functions can be accumulated.
A first function was added to overcome a drawback of gtest. Here one could not
compare the result of any of the SageBuilder::buildTypeXYZ with NULL, since int and
SgXYZType* are non-comparable datatypes.
The new function uses a dynamic cast and returns bool.