Delayed the issue, why a lot of unit tests for function types failed.
I realized, that the global type table is a global datastructure, that isn't cleaned after
each test run, so it gets polluted over time.
I wrote a test fixture, that erases every entry of that typetable - in an inefficient way
right now. However.
The source of the problem seems to be more tricky.