(Binary Analysis) New capability for BaseSemantics::RegisterStateGeneric
The coalesceOnRead Boolean property controls whether readRegister adjusts
the representation of existing registers when the register being read overlaps
with registers that are already stored.  The two choices are:

1. Adjust existing representation so that the just-read value can be
   stored atomically.

2. Don't adjust existing representation, but rather fit parts of the
   just-read value into overlapping registers and the gaps between them.

Also the x86 dispatcher was modified so that coalesceOnRead is disabled
whenever its reading FLAGS or EFLAGS. This will cause flag bits, that
analysts tend to think of individually, to stay individual bits even
when the whole register is read.