(Boost) Add include <path> to EDG and backend CLI (--with-boost=<path>)
This addresses an issue reported by Rice University. My error report
to them after investigation:

```
It looks like some of the errors are related to a possible mis-match in
Boost versions.

For example,

    test2013_241.C [err]: /usr/include/boost/mpl/apply.hpp:138:31: error:
missing binary operator before token "("

This test is using "/usr/include/boost", however ROSE was configured to
use Boost from "/projects/pkgs/boost".
```

The problem is that the ROSE/configure --with-boost=<path> is not
being added to the EDG or backend compiler commandlines. It's peculiar
that this ever worked because it turns out that Rice's /usr/include
version of Boost is 1.33.1, which is the same for our RHEL5 machines
in Jenkins.

The mysteries of software development continues...