Use CMake for more robust compiler detection
When building with CMake, we can take advantage of some of the
variables that it automatically sets to detect whether or not
we are building ROSE with GNU compilers.  This is a more robust
solution than matching based on the name of the compiler executable,
which can vary from system to system.