(Binary Analysis) Added switch to control whether functions return
The "--functions-return=BOOLEAN" switch allows the user to specify
whether functions for which the may-return analysis is indeterminate
should be assumed to return or not return to their caller.  The
default is to assume they return; the alternative is mainly to
test that things are working properly by comparing two disassembly
listings.