(Binary Analysis) Intra-function stack delta analysis
Computes stack deltas at the beginning and end of each basic
block with respect to the stack pointer at the beginning of
the function. See Partitioner::basicBlockStackDeltaIn,
basicBlockStackDeltaOut, and functionStackDelta.

TODO: Cache the stack delta results in the AST. Currently the
results are only available inside Partitioner2.

TODO: Mechanism for handling callee-cleanup common in PE executables