(Binary Analysis) Tweaks to the order that blocks are discovered
The partitioner is slightly more efficient if it can discover basic
blocks in depth-first order since many of the analysis phases are
also depth-first.

The may-return worklist is sorted in reverse depth-first order
but this code is currently disabled because it causes the partitioner
to be about 10x slower (optimized) than with the naive approach.