(Binary Analysis) More special cases for may-return by definition
These cases are to help speed up the analysis by deciding on a
may-return analysis result earlier in the process rather than
later. They assume that certain types of CFG connectivity are
unlikely to change, such as a basic block whose address is not
available, or a basic block whose only successor is indeterminate.