(Binary Analysis) Speed improvements for adding call-ret edges
Use three lists of basic blocks that might need a call-return edge
to be added so that we can delay sorting the final list as long
as possible and therefore do it in as large a batch as possible. The
main effective change is that may-return analysis is performed at
most twice per basic block, but the results seem to be just as
good as the slower method where I've compared them.