(Support) configure switch --with-yaml to add YAML support
We had a request long ago to be able to read configuration information
for binary analysis from JSON files. This commit is the first step in
that direction.

The JSON format is a subset of YAML and the latter has some useful
extra features (like comments in the config files).  The libyaml-cpp
is a C++ API using C++ paradigms (i.e., much more than just a wrapper
around a C-like API) and seems to be stable, and is part of many
GNU/Linux distributions.

The --with-yaml is entirely optional.