(Binary Analysis) Ability to label addresses
The user can add a label to any address and the partitioner
will automatically use that label as function names or as comments
on constants.

The Engine has a method that will add address labels for all
symbols in the ELF/PE container.

Added a function that will add comments to addresses of NUL-terminated
ASCII strings.