(Binary Analysis) bROwSE: major reorganization of app-level stuff
Reorganized Application and WFunction in order to support configuration
inside the browser rather than the command line.

The HexDump window is temporarily gone, but will be back.

Adds support for re-running the partitioner with different
configuration, although the page could be improved some.

Tabbed windows are enabled/disabled based on context.

Memory map can be editable or not, depending on context.