SMTPathFeasibility project added
Fixed an issue with mod operator used by user in pragma, z3 tests now working, but need
to be set up so they only run if the user has z3

Adding make check tests to Makefile.am

Fixed issues that caused ROSE to not be able to check for the existence of the Z3 binary

Combining original pragma functionality and current Result Query functionality

In progress, working on checking order of operations is correct and adding functions
in SMT2 to work with C's bool = 0, 1

Moved all tests into passing/failing groups, reintegrated the SMT2 solver without Result/Query.
Began work on documentation in the docs subfolder.

Added choice for z3 support, and the makefile will allow z3 tests if z3
is enabled