(Binary Analysis) Multi-byte memory I/O operations are more stringent
Memory states are always constructed with unspecified byte order (no change
in behavior) because their construction usually occurs before an architecture
is determined.

Previously, multi-byte readMemory or writeMemory RISC operations assumed
that memory was little endian if its byte order was not specified. New behavior
is to throw a semantics exception in this case.

To alleviate having to define a byte order for every memory state, the semantics
dispatcher (e.g., DispatcherX86) constructors now warn if the byte order has an
incompatible specified value, or set the byte order if unspecified.