(Binary Analysis) Fixed bug in MemoryCellList
MemoryCellList::readMemory had a bug when reading from a memory
state that had a single cell.  If the single cell's address was
a possible alias (may_alias) with the address being read, then
readMemory returned the value stored in that cell.  New behavior
is to return the merge of the cell's value with the "dflt" value
provided in the readMemory call.

Test case:
   mov dword [eax], 0;
   mov ebx, [ecx];

Before the fix, the EBX final value is zero.  After the fix,
the EBX final value is unknown.