Removed "subdir-objects" from automake
Dan added this to configure.ac when he renamed configure.in (he deleted
configure.in in f50cdbc6 and added configure.ac in 6cfbaf61), and it
turns off lots of warnings. However, it causes automake 1.11.1 (and
possibely others) to generate makefiles which, when "make" runs,
modify the $ROSE_SRC tree.  This modification causes parallel building
to have race conditions.

Therefore, I'm reverting this change by removing "subdir-objects".
We may need to require a newer version of automake, or rewrite
some makefiles so they don't require subdir-objects. But this whole
problem will go away if we move to only cmake.