(Binary Analysis) Made more of SymbolicExprParser public
This commit makes the lexical analyzer public so that the expression
parser can be invoked more than once on a single stream of input,
and/or the user can have custom parsing.

The findPath tool uses this to parse register and memory post
conditions of the form "regname = expression" and "address = expression".