(Binary Analysis) Abstracted functions in expression parser
Functions are now handled via function table, allowing the user
to provide their own functions that expand into symbolic expressions.
The default parser now uses this feature to parse SMT and C-like function
names:

  input> (+ 1 2)

The function is "+", which gets looked up in the function table,
a C++ functor is invoked to create a symbolic expression whose
operator is the enum OP_ADD and whose arguments are the same
as for the "+".