(Binary Analysis) Pointer detection second iteration
This iteration does better tracking between pointer l- and r-values
resulting in fewer false positives.

Tests are not updated yet because this iteration goes a little two far,
not traversing backward through temporary locations to find the
original pointer. E.g., it finds [ebp-4] but not [ebp+8] in this
code:

   mov eax, [ebp+8]
   ...
   mov [ebp-4], eax ; spilled
   ...
   mov eax, [ebp-4] ; restored
   mov eax, [eax]   ; dereferenced