(Support) The "--threads=N" argument stores N in a global variable.
Created a global variable (CommandlineProcessing::genericSwitchArgs) for
storing the arguments for generic switches. Previous behavior required the
user to query the command-line parser result to get the arguments, but the
result was not always available if the parsing was done deeper in ROSE.

This isn't an ideal solution because it assumes that only one command-line
is parsed. But that's usually the case, so it works fine most of the time.