(Binary Analysis) whole-program calling convention is now parallel
Uses the new Sawyer::workInParallel to run calling-convention
analysis on all functions, invoking it in callees before callers
except when there are cycles in the call graph. The analysis is
currently slow, so it has to be enabled with "--post-calling-convention"
the number of worker threads is controled with "--threads=N". It's
been tested with up to 72 threads using 7 hours of CPU time.