(Attributes) Revert "Turned off emulation of LEAK_VALUES_BUG"
It turns out that the Java Fortran parser depends on this bug: the
attribute mechanism is not allowed to take ownership of the heap-
allocated attributes and delete them when they're no longer needed
by the attribute mechanism.

Added commentary to the #ifdef line that causes this bug emulation
to be enabled.