(Binary Analysis) Added trimming mechanism to SymbolicSemantics
The RiscOperators::trimThreshold property imposes a maximum size
on symbolic expressions (the default is zero: no size limit). If
an expression becomes larger than this limit then it's replaced
by a new variable.

The filterResult virtual method is also a conventient function
for users to augment if they need to do something with every result.