(Binary Analysis) GraphViz node names are mostly virtual addresses
When a GraphViz file is generated, the nodes of the graph were
numbered based on the order they were emitted (their labels had
info about their addresses, instructions, etc).  However, we want
to be able to use "dot" to process a graph and generate positions
for the layout, and then read that information back into ROSE and
attach it to the graph.  Therefore, the GraphViz nodes are now
given names which are mostly hexadecimal strings representing the
virtual address of the corresponding CFG vertex.