(Binary Analysis) Partitioner allows shared basic blocks
A basic block can now be owned by multiple functions.  This can
lead to ambiguous classification of whether an edge is inter-
or intra-procedural, and other details that need to be worked
out still.

An AST is not yet supported since sharing implies that the
AST might not be a tree.