Qualified "boost::tie" in graphProcessing.h
If we really want these algorithms to be generic (as the use of
templates implies) then they should qualify all the BGL functions
because Koenig lookup will fail if the graph implementation isn't
in the boost namespace. I've run into this before with BGL, but
I didn't change this file any more than necessary.