(Support) SqlDatabase API functions for parsing URLs
I've exposed the functions that ROSE uses for parsing database
URLs because they're useful when the user wants to use some other
database library.  It allows the other tool to specify database
names in the same way as the rest of ROSE even though the underlying
database libraries don't generally accept a URL style of naming.