(Binary Analysis) Fixed lack of edges in BinaryAnalysis::FunctionCall::Graph
The build_cg_from_ast function failed to insert any edges into the
function call graph when the AST was created by Partitioner2 because
the basic block successors for call blocks, which are of type
SgAsmIntegerValue, have base objects that are SgAsmFunction instead
of SgAsmBlock. I.e., calls point to functions instead of the first
block of a function.

ROSE-287