(FIX) unparse attachedPreprocessingInfo inside the body of a Fortran ELSE stmt.
IF (test) THEN
!$ comment1
ELSE IF (test) THEN
!$ comment2
END IF

Comment2 was not unparsed because it was attached to the false body of the IF stmt.
The true body of the ELSEIF stmt was unparsed, and the false body of the IF stmt was NOT unparsed.
This could be a problem if only a comment was inside the false body of the first IF stmt.