(Configure) Run the compiler to find version #define's
We need to #define things like __GNUC__, __GNUC_MINOR__, and
__GNUC_PATCH_LEVEL__ for g++, clang++, and icpc (yes, they all
define these macros, checked by Dan).  Previously, we obtained
values for these macros by running the compiler with "--version"
or "-dumpversion", but now we run the compiler with "-dM". This
works better because "g++ -dumpversion" doesn't always emit the
patch level, in which case the old code was adding
"-D__GNUC_PATCH_LEVEL__=0" to the EDG commandline, which was wrong.