(Util) Added "using namespace rose" to some library .C files
Those files that are part of the ROSE library but whose contents
are not in the "rose" namespace should have a "using namespace rose".
Eventually all (most?) of these will be migrated into "rose" and will
no longer need the "using namespace rose".

This was part of our discussion to productizing ROSE -- a library
should live inside a namespace so as to not pollute the user's
space.

In certain places (mostly headers and non-C files) I chose to
qualify the names instead of adding "using namespace rose".