(Utility) Added "using namespace rose" to the dataStructureGraphing project
As the ROSE library declarations move into the "rose" namespace, as
with ::StringUtility -> ::rose::StringUtility in a recent commit,
users of librose (projects, tests, etc) will need to either qualify
the ROSE functions and types or add a "using namespace rose". Thus
they'll do the same thing with librose as they do with the STL, boost,
etc.

I've added "using namespace rose" to this project because it
uses rose::StringUtility.