(Utility) Added "using namespace rose" to tests that use the ROSE library
Since tests are like users, they need to either qualify the ROSE functions
with rose:: or have a "using namespace rose". This will become more
important as the ROSE library migrates into the "rose" namespace.

The changes in this commit are to fix compile errors since ::StringUtility
was renamed to ::rose::StringUtility.